asset+commitment: properly merge asset commitments
This commit fixes an issue with the NewTapCommitment function which used
to overwrite asset commitments if they had the same Taproot Asset
commitment key (which is true for grouped assets for example).

+++ asset/mock.go
@@ -358,7 +358,7 @@ func RandAssetWithValues(t testing.TB, genesis Genesis, groupKey *GroupKey,
 
 	t.Helper()
 
-	units := test.RandInt[uint64]() + 1
+	units := test.RandInt[uint32]() + 1
 
 	switch genesis.Type {
 	case Normal:
@@ -376,7 +376,7 @@ func RandAssetWithValues(t testing.TB, genesis Genesis, groupKey *GroupKey,
 	}
 
 	a, err := New(
-		genesis, units, 0, 0, scriptKey, groupKey,
+		genesis, uint64(units), 0, 0, scriptKey, groupKey,
 		WithAssetVersion(assetVersion),
 	)
 	require.NoError(t, err)
